feat: Chart optionally installs CRD / CR Manager configurable for more strict clusters (#344)
Coupling the CRD creation to a controller requires an unnecessarily privileged cluster role.

In some clusters, admins would not be willing to grant the controller service account that role. We require the chart itself to install the CRD, and the service account or admin installing the chart would have the crd:create access while the controller be more tightly scoped.

This essentially deprecates the custom resource manager so that it can removed entirely in future 4.x releases.