fix: Additional quotes added while creating secret
As the JSON.stringify adds additional quotes when it reads the json data from the vault and while creating the secret object in K8s it adds additional quotes to avoid it just we need to make sure it is parsed to remove it and have only the actual value used for the secret in K8s. I have made this change in my usage and it is working perfectly fine.