Add 'dataFromLiteral' to ExternalSecret API
I have a use case where I need a Kubernetes secret which contains a file
meant to be mounted to a container. In this file there is sensitive data
so it must be templated using the `spec.template.stringData`
functionality. Within this file there is also recurring, non-sensitive,
data that would benefit from being templated.

More specifically, I had a dozen references to a GCP project ID which
would need to be hard coded in that many locations.

I tried to supply this within the `spec.template.stringData`, but
expectedly, that data is not part of the map which renders the template;
only `spec.data` and `spec.dataFrom` are merged for that.

It would be possible to use Helm to fill Golang templates within the
lodash template, however, if feels like a bad practice to have 2
templating patterns operating on the same data.

This led me to introduce the `dataFromLiteral` key-value pairing for the
abstract kvbackend so all backends can template non-sensitive data along
side their sensitive, cloud secrets.